// src/app/interface/actorsList-interface.ts
export interface ActorList{
    id: number;
    actor_id: number;
    actor_firstname: string;
    actor_lastname: string;
    actor_job: string;
    actor_comment: string;
    actor_At11: string;
    actor_At12: string;
    actor_At13: string;
    actor_At14: string;
    actor_At21: string;
    actor_At22: string;
    actor_At23: string;
    actor_At31: string;
    actor_At32: string;
    actor_At33: string;
    actor_At41: string;
    actor_At42: string;
    actor_At51: string;
    actor_At52: string;
    actor_At53: string;
    
}